pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "quotation-intelligence"
version = "0.1.0"
description = "SaaS platform for converting PDF quotations to structured data"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.6.0"
sqlalchemy = {extras = ["asyncpg"], version = "^2.0.36"}
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.2.0"
pdfplumber = "^0.11.4"
pymupdf = "^1.24.14"
pytesseract = "^0.3.13"
pillow = "^11.0.0"
litellm = "^1.40.0"
python-multipart = "^0.0.17"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
structlog = "^24.4.0"
openpyxl = "^3.1.5"
pandas = "^2.2.3"
tenacity = "^9.0.0"
boto3 = "^1.35.0"
httpx = "^0.27.0"
prometheus-client = "^0.21.0"
sentry-sdk = {extras = ["fastapi"], version = "^2.19.0"}
asyncpg = "^0.31.0"
ipykernel = "^7.2.0"
requests = "^2.33.1"
pymupdf4llm = "^1.27.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
httpx = "^0.27.0"
black = "^24.10.0"
ruff = "^0.8.0"
mypy = "^1.13.0"
pre-commit = "^4.0.1"
factory-boy = "^3.3.1"
freezegun = "^1.5.1"
responses = "^0.25.3"

[tool.poetry.scripts]
quotation-api = "quotation_extraction.api.main:main"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E203", "E501", "W505"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=quotation_extraction --cov-report=term-missing"

[tool.coverage.run]
source = ["quotation_extraction"]
omit = ["*/tests/*", "*/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
